backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
PyMySQL==1.1.0
PyJWT==2.8.0
bcrypt==4.1.2
python-dotenv==1.0.0
gunicorn==21.2.0

# OpenAI for transcription and analysis
openai==1.6.1

# Audio processing
pydub==0.25.1
mutagen==1.47.0

# PDF generation
reportlab==4.0.7

# Validation
validators==0.22.0

# Google OAuth
google-auth==2.41.1
google-auth-oauthlib==1.2.3
google-auth-httplib2==0.3.0
